requirements.txt
Django==2.2
django-single-instance-model==1.1.0
django-taggit==1.2.0
pytz==2019.3
sqlparse==0.3.1
